+++ awint/src/lib.rs
@@ -7,7 +7,7 @@ pub use awint_core::{bw, Bits, InlAwi, SerdeError};
 #[cfg(feature = "dag")]
 pub use awint_dag;
 #[cfg(feature = "alloc")]
-pub use awint_ext::ExtAwi;
+pub use awint_ext::{ExtAwi, FPType, FP};
 pub use awint_macros::*;
 
 /// Reexports every user-intended macro, structure, and function except for
@@ -17,7 +17,7 @@ pub mod prelude {
 
     #[cfg(feature = "alloc")]
     pub use crate::ExtAwi;
-    pub use crate::{bw, cc, Bits, InlAwi};
+    pub use crate::{bw, cc, Bits, FPType, InlAwi, FP};
 }
 
 #[cfg(feature = "awint_dag")]